Push down code related to java partition scanners, now that o.e.jface.text.rules have been moved to org.eclipse.text
