+++ org.eclipse.jdt.core.manipulation/common/org/eclipse/jdt/internal/ui/text/FastJavaPartitioner.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2020 IBM Corporation and others.
+ * Copyright (c) 2019, 2023 IBM Corporation and others.
  *
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
@@ -74,8 +74,8 @@ public class FastJavaPartitioner extends FastPartitioner {
 
 	private boolean isTextBlockSupported() {
 		boolean isTextBlockSupported= false;
-		if (fScanner instanceof FastJavaPartitionScanner) {
-			isTextBlockSupported= ((FastJavaPartitionScanner) fScanner).isTextBlockSupported();
+		if (fScanner instanceof AbstractFastJavaPartitionScanner) {
+			isTextBlockSupported= ((AbstractFastJavaPartitionScanner) fScanner).isTextBlockSupported();
 		} else {
 			isTextBlockSupported= false;
 		}
